+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AABB` and `IArgument` extension methods
 - ItemStack lua formatting
 - Interfaces for upgrade that contains other upgrades. Useful for different render tricks and other staff
+- Validation logic for loot table generators
 
 ### Fixed